+++ libs/syntax/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Handle `underscoreUpper` for names (Example: FIRST_NAME).
 - Handle `hyphenUpper` for names (Example: FIRST-NAME).
 - Handle subtype `rich` of type `string`.
+- Handle `isGeoSearchable` property.
 
 ### Changed
 - Use of `safe-eval` for evaluation